Fixed some circular dependecies
