+++ packages/ts-components/src/components/article-flag/types.ts
@@ -1,0 +1,4 @@
+export type FlagType = Array<{
+  expiryTime: string | null;
+  type: string;
+}>;
