+++ packages/ts-components/src/components/in-article-info-card/types.ts
@@ -1,0 +1,8 @@
+export type InfoCardData = {
+  type: string;
+  data: {
+    image?: string;
+    subtitle?: string;
+    copy: string;
+  };
+};
