+++ packages/ts-components/src/components/save-star/ContentProvider.tsx
@@ -1,12 +1,6 @@
 import React from 'react';
 import { useFetch } from '../../helpers/fetch/FetchProvider';
-import { ArticleBookmark } from './SaveStarUI';
-
-export interface ContentProps {
-  loading?: boolean;
-  error?: string;
-  data?: ArticleBookmark;
-}
+import { ArticleBookmark, ContentProps } from './types';
 
 export const ContentProvider: React.FC = React.memo(({ children }) => {
   const fetchResponse = useFetch<ArticleBookmark>();
