+++ packages/ts-components/src/components/save-star/types.ts
@@ -1,0 +1,9 @@
+export type ArticleBookmark = {
+  isBookmarked: boolean;
+};
+
+export interface ContentProps {
+  loading?: boolean;
+  error?: string;
+  data?: ArticleBookmark;
+}
